+++ src/data/changelog-version.ts
@@ -1,3 +1,3 @@
 export function getLatestVersion(): { version: string; date: string } {
-  return { version: "8.94.3", date: "2026-03-20" };
+  return { version: "8.94.4", date: "2026-03-20" };
 }